requirements.txt
flask
flask-cors
scikit-learn
numpy
pandas
pymongo
python-dotenv
